% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(x, conv = FALSE)
}
\arguments{
\item{x}{results from the fitted model.}

\item{conv}{defulated for odds ratio plot, otherwise show the iteration plot.}
}
\value{
plot odds ratio with CIs or plot of the iterations.
}
\description{
This function gives plot of the odds ratio or shows the iteration for convergence.
}
\examples{
 beta=c(0.1,0.2,0.3)
 alpha=c(0.4,0.5)
 variance=c(0.8,0.5,0.7,0.3,0.4)
 x=list(beta,alpha,variance)
 names(x)=c("beta","alpha","variance")
 plot_model(x)
}
