% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mgee2.R
\name{print.summary.mgee2}
\alias{print.summary.mgee2}
\title{print.summary.mgee2}
\usage{
\method{print}{summary.mgee2}(x, ...)
}
\arguments{
\item{x}{the summary results}

\item{...}{Other parameters}
}
\value{
a table of summary statistics
}
\description{
print.summary.mgee2
}
