% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mgee2.R
\name{summary.mgee2}
\alias{summary.mgee2}
\title{summary.mgee2}
\usage{
\method{summary}{mgee2}(object, ...)
}
\arguments{
\item{object}{The fitted model}

\item{...}{Other parameters
summary function for mgee2 method output}
}
\description{
summary.mgee2
}
